requirements.txt
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
Flask==2.2.2
Flask-Session==0.4.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.3.3
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
